svn获取历史记录和清单申请展示

+++ src/main/java/org/tis/senior/module/developer/service/ISSvnKitService.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface ISSvnKitService {
 
-    List<SvnCommit> loadSvnHistory(String url, Long startRevision);
+    List<SvnCommit> loadSvnHistory(String url, int startRevision);
 }
